requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
config==0.5.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
pipenv==2022.6.7
pkg_resources==0.0.0
platformdirs==2.5.2
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.15.1
virtualenv-clone==0.5.7
whitenoise==6.2.0
